+++ assets/javascripts/kitten/components/dropdowns/user-menu.js
@@ -1,5 +1,6 @@
-import React from 'react'
+import React, { Component } from 'react'
 import PropTypes from 'prop-types'
+import classNames from 'classnames'
 import { Dropdown } from 'kitten/components/dropdowns/dropdown'
 import { ButtonImageWithTextAndBadge } from
   'kitten/components/buttons/button-image-with-text-and-badge'
@@ -8,7 +9,7 @@ import domElementHelper from 'kitten/helpers/dom/element-helper'
 import emitter from 'kitten/helpers/utils/emitter'
 import objectAssign from 'core-js/library/fn/object/assign'
 
-export class UserMenu extends React.Component {
+export class UserMenu extends Component {
   constructor(props) {
     super(props)
 
@@ -123,7 +124,8 @@ export class UserMenu extends React.Component {
 
   getDropdownContent() {
     return (
-      <NavList className="k-UserMenu k-UserMenu--withoutBorderTop"
+      <NavList className={ classNames('k-UserMenu',
+                                      'k-UserMenu--withoutBorderTop') }
                role="menubar"
                list={ this.props.dropdownList } />
     )