+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking change: Update `UserMenu` component with Kitten V2.
+  Replace `@include k-UserMenu(( font: 'bold', … ))`
+  by `@include k-UserMenu;`.
 - Feature: Update `HorizontalNav` component with Kitten V2.
 - Feature: Add `tiny` props in `TagButton` component.
 - Feature: Update `TagButton` and `PaymentButton` components with Kitten V2.